fix sync not handeling updating sg, st, unit codes
