+++ src/CruiseDAL.V3.Sync.Test/CruiseSyncer_Test.cs
@@ -7,7 +7,7 @@ using Xunit.Abstractions;
 
 namespace CruiseDAL.V3.Sync
 {
-    public class CruiseSyncer_Test : Datastore_TestBase
+    public partial class CruiseSyncer_Test : Datastore_TestBase
     {
         public CruiseSyncer_Test(ITestOutputHelper output) : base(output)
         {
@@ -226,141 +226,6 @@ namespace CruiseDAL.V3.Sync
         }
 
         [Fact]
-        public void Sync_Unit_Add()
-        {
-            var fromPath = base.GetTempFilePath(".crz3", "Sync_Unit_Add_fromFile");
-            var toPath = base.GetTempFilePath(".crz3", "Sync_Unit_Add_toFile");
-
-            var syncOptions = new CruiseSyncOptions();
-
-            var cruiseID = Guid.NewGuid().ToString();
-            var saleID = Guid.NewGuid().ToString();
-
-            using var fromDb = CreateDatabaseFile(fromPath, cruiseID, saleID);
-
-            fromDb.CopyTo(toPath, true);
-            using var toDb = new CruiseDatastore_V3(toPath);
-
-            var newUnit = new CuttingUnit()
-            {
-                CruiseID = cruiseID,
-                CuttingUnitID = Guid.NewGuid().ToString(),
-                CuttingUnitCode = "newUnitCode1",
-            };
-            fromDb.Insert(newUnit);
-
-            var syncer = new CruiseSyncer();
-            syncer.Sync(cruiseID, fromDb, toDb, syncOptions);
-
-            toDb.GetRowCount("CuttingUnit", "WHERE CuttingUnitID = @p1", newUnit.CuttingUnitID)
-                .Should().Be(1);
-        }
-
-        [Fact]
-        public void Sync_Unit_Update()
-        {
-            var fromPath = base.GetTempFilePath(".crz3", "fromFile");
-            var toPath = base.GetTempFilePath(".crz3", "toFile");
-
-            var syncOptions = new CruiseSyncOptions();
-
-            var cruiseID = Guid.NewGuid().ToString();
-            var saleID = Guid.NewGuid().ToString();
-
-            using var fromDb = CreateDatabaseFile(fromPath, cruiseID, saleID);
-
-            fromDb.CopyTo(toPath, true);
-            using var toDb = new CruiseDatastore_V3(toPath);
-
-            var unit = fromDb.From<CuttingUnit>().Query().First();
-            unit.Area = Rand.Int();
-            unit.Description = Rand.String();
-            unit.LoggingMethod = "401";
-            unit.PaymentUnit = Rand.AlphaNumeric(3);
-            unit.Rx = Rand.AlphaNumeric(3);
-            unit.ModifiedBy = Rand.AlphaNumeric(4);
-            fromDb.Update(unit);
-
-            var syncer = new CruiseSyncer();
-            syncer.Sync(cruiseID, fromDb, toDb, syncOptions);
-
-            var unitAgain = toDb.From<CuttingUnit>().Where("CuttingUnitID = @p1").Query(unit.CuttingUnitID).First();
-
-            unitAgain.Should().BeEquivalentTo(unit, x => x.Excluding(y => y.Modified_TS));
-        }
-
-        [Fact]
-        public void Sync_Stratum_Add()
-        {
-            var fromPath = base.GetTempFilePath(".crz3", "Stratum_Add_fromFile");
-            var toPath = base.GetTempFilePath(".crz3", "Stratum_Add_toFile");
-
-            var syncOptions = new CruiseSyncOptions();
-
-            var cruiseID = Guid.NewGuid().ToString();
-            var saleID = Guid.NewGuid().ToString();
-
-            using var fromDb = CreateDatabaseFile(fromPath, cruiseID, saleID);
-
-            fromDb.CopyTo(toPath, true);
-            using var toDb = new CruiseDatastore_V3(toPath);
-
-            var stratumID = Guid.NewGuid().ToString();
-            fromDb.Insert(new Stratum()
-            {
-                CruiseID = cruiseID,
-                StratumID = stratumID,
-                StratumCode = "10",
-                Method = "100",
-            });
-            var newStratum = fromDb.From<Stratum>().Where("StratumID = @p1").Query(stratumID).Single();
-
-            var syncer = new CruiseSyncer();
-            syncer.Sync(cruiseID, fromDb, toDb, syncOptions);
-
-            var stratumAgain = toDb.From<Stratum>().Where("StratumID =  @p1")
-                .Query(stratumID).FirstOrDefault();
-            stratumAgain.Should().BeEquivalentTo(newStratum, x => x.Excluding(y => y.Modified_TS));
-        }
-
-        [Fact]
-        public void Sync_Stratum_Update()
-        {
-            var fromPath = base.GetTempFilePath(".crz3", "Stratum_Updated_fromFile");
-            var toPath = base.GetTempFilePath(".crz3", "Stratum_Update_toFile");
-
-            var syncOptions = new CruiseSyncOptions();
-
-            var cruiseID = Guid.NewGuid().ToString();
-            var saleID = Guid.NewGuid().ToString();
-
-            using var fromDb = CreateDatabaseFile(fromPath, cruiseID, saleID);
-
-            var stratumID = Guid.NewGuid().ToString();
-            var stratum = new Stratum()
-            {
-                CruiseID = cruiseID,
-                StratumID = stratumID,
-                StratumCode = "10",
-                Method = "100",
-            };
-            fromDb.Insert(stratum);
-
-            fromDb.CopyTo(toPath, true);
-            using var toDb = new CruiseDatastore_V3(toPath);
-
-            stratum.Hotkey = Rand.String();
-            fromDb.Update(stratum);
-
-            var syncer = new CruiseSyncer();
-            syncer.Sync(cruiseID, fromDb, toDb, syncOptions);
-
-            var stratumAgain = toDb.From<Stratum>().Where("StratumID =  @p1")
-                .Query(stratumID).FirstOrDefault();
-            stratumAgain.Should().BeEquivalentTo(stratum, x => x.Excluding(y => y.Modified_TS));
-        }
-
-        [Fact]
         public void Sync_UnitStratum_Add()
         {
             var fromPath = base.GetTempFilePath(".crz3", "fromFile");
@@ -391,77 +256,6 @@ namespace CruiseDAL.V3.Sync
 
             toDb.GetRowCount("CuttingUnit_Stratum", "WHERE CuttingUnitCode = @p1 AND StratumCode = @p2", newCUST.CuttingUnitCode, newCUST.StratumCode)
                 .Should().Be(1);
-        }
-
-        [Fact]
-        public void Sync_SampleGroup_Add()
-        {
-            var fromPath = base.GetTempFilePath(".crz3", "SampleGroup_Add_fromFile");
-            var toPath = base.GetTempFilePath(".crz3", "SampleGroup_Add_toFile");
-
-            var syncOptions = new CruiseSyncOptions();
-
-            var cruiseID = Guid.NewGuid().ToString();
-            var saleID = Guid.NewGuid().ToString();
-
-            using var fromDb = CreateDatabaseFile(fromPath, cruiseID, saleID);
-
-            fromDb.CopyTo(toPath, true);
-            using var toDb = new CruiseDatastore_V3(toPath);
-
-            var sampleGroupID = Guid.NewGuid().ToString();
-            fromDb.Insert(new SampleGroup()
-            {
-                CruiseID = cruiseID,
-                SampleGroupID = sampleGroupID,
-                SampleGroupCode = "10",
-                StratumCode = Strata[0].StratumCode,
-            });
-            var newSampleGroup = fromDb.From<SampleGroup>().Where("SampleGroupID = @p1").Query(sampleGroupID).Single();
-
-            var syncer = new CruiseSyncer();
-            syncer.Sync(cruiseID, fromDb, toDb, syncOptions);
-
-            var sampleGroupAgain = toDb.From<SampleGroup>().Where("SampleGroupID =  @p1")
-                .Query(sampleGroupID).FirstOrDefault();
-            sampleGroupAgain.Should().BeEquivalentTo(newSampleGroup);
-        }
-
-        [Fact]
-        public void Sync_SampleGroup_Update()
-        {
-            var fromPath = base.GetTempFilePath(".crz3", "SampleGroup_Update_fromFile");
-            var toPath = base.GetTempFilePath(".crz3", "SampleGroup_Update_toFile");
-
-            var syncOptions = new CruiseSyncOptions();
-
-            var cruiseID = Guid.NewGuid().ToString();
-            var saleID = Guid.NewGuid().ToString();
-
-            using var fromDb = CreateDatabaseFile(fromPath, cruiseID, saleID);
-
-            var sampleGroupID = Guid.NewGuid().ToString();
-            var sampleGroup = new SampleGroup()
-            {
-                CruiseID = cruiseID,
-                SampleGroupID = sampleGroupID,
-                SampleGroupCode = "10",
-                StratumCode = Strata[0].StratumCode,
-            };
-            fromDb.Insert(sampleGroup);
-
-            fromDb.CopyTo(toPath, true);
-            using var toDb = new CruiseDatastore_V3(toPath);
-
-            sampleGroup.Description = Rand.String();
-            fromDb.Update(sampleGroup);
-
-            var syncer = new CruiseSyncer();
-            syncer.Sync(cruiseID, fromDb, toDb, syncOptions);
-
-            var sampleGroupAgain = toDb.From<SampleGroup>().Where("SampleGroupID =  @p1")
-                .Query(sampleGroupID).FirstOrDefault();
-            sampleGroupAgain.Should().BeEquivalentTo(sampleGroup, x => x.Excluding(y => y.Modified_TS));
         }
 
         [Fact]
@@ -505,87 +299,9 @@ namespace CruiseDAL.V3.Sync
                 .Excluding(y => y.CreatedBy));
         }
 
-        [Fact]
-        public void Sync_Plot_Add()
-        {
-            var fromPath = base.GetTempFilePath(".crz3", "Sync_Plot_Add_fromFile");
-            var toPath = base.GetTempFilePath(".crz3", "Sync_Plot_Add_toFile");
+        
 
-            var syncOptions = new CruiseSyncOptions()
-            {
-                Design = SyncFlags.Insert,
-                FieldData = SyncFlags.Insert,
-            };
-
-            var cruiseID = Guid.NewGuid().ToString();
-            var saleID = Guid.NewGuid().ToString();
-
-            using var fromDb = CreateDatabaseFile(fromPath, cruiseID, saleID);
-
-            fromDb.CopyTo(toPath, true);
-            using var toDb = new CruiseDatastore_V3(toPath);
-
-            var newPlot = new Plot()
-            {
-                CruiseID = cruiseID,
-                PlotID = Guid.NewGuid().ToString(),
-                CuttingUnitCode = Units[0],
-                PlotNumber = 1,
-            };
-            fromDb.Insert(newPlot);
-
-            var syncer = new CruiseSyncer();
-            syncer.Sync(cruiseID, fromDb, toDb, syncOptions);
-
-            toDb.GetRowCount("Plot", "WHERE PlotID = @p1", newPlot.PlotID)
-                .Should().Be(1);
-        }
-
-        [Fact]
-        public void Sync_Plot_Update()
-        {
-            var fromPath = base.GetTempFilePath(".crz3", "Sync_Plot_Add_fromFile");
-            var toPath = base.GetTempFilePath(".crz3", "Sync_Plot_Add_toFile");
-
-            var syncOptions = new CruiseSyncOptions()
-            {
-                Design = SyncFlags.Insert,
-                FieldData = SyncFlags.InsertUpdate,
-            };
-
-            var cruiseID = Guid.NewGuid().ToString();
-            var saleID = Guid.NewGuid().ToString();
-
-            using var fromDb = CreateDatabaseFile(fromPath, cruiseID, saleID);
-
-            var plot = new Plot()
-            {
-                CruiseID = cruiseID,
-                PlotID = Guid.NewGuid().ToString(),
-                CuttingUnitCode = Units[0],
-                PlotNumber = 1,
-            };
-            fromDb.Insert(plot);
-
-            fromDb.CopyTo(toPath, true);
-            using var toDb = new CruiseDatastore_V3(toPath);
-
-            plot.Remarks = Rand.String();
-            plot.Slope = Rand.Double();
-            plot.PlotNumber = Rand.Int();
-            fromDb.Update(plot);
-
-            var syncer = new CruiseSyncer();
-            syncer.Sync(cruiseID, fromDb, toDb, syncOptions);
-
-            var plotAgian = toDb
-                .From<Plot>()
-                .Where("PlotID = @p1")
-                .Query(plot.PlotID)
-                .FirstOrDefault();
-
-            plotAgian.Should().BeEquivalentTo(plot, config => config.Excluding(x => x.Modified_TS));
-        }
+        
 
         [Fact]
         public void Sync_Plot_Stratum_Add()
